Add settings option to use rich traceback
This comes from a personal use that I thought might help others.
I had many GB of data in RAM when a crash happened in brainrender.
Rich tried to display my local variables to terminal, and couldn't.
Disabling rich printing is best here.

There may be a better solution to this problem, but since the
previous code loading the rich style traceback at init of brainrender
this was the best I could think of.